+++ CHANGELOG.md
@@ -42,4 +42,13 @@ All notable changes to this project will be documented in this file, following t
  - Added package info for the codec project
  - Improved docs and refactoring of test names
 
+## v0.2.2 - 2016-07-05
+### Changed
+ - Refactored the generation of Bioassemblies  
+ - Refactored generateGroupMap to generateGroupList
+ - Update to the serialization module - only construct object mapper once
+
+### Added 
+ - getBioassemblyName added to the API - return the BioassemblyName as a string.
+ - Tests for EncoderUtils